remove fips disabled checks from non-fips tests
