+++ transport/http_test.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -204,26 +203,6 @@ func TestHTTPTransportTLS(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestHTTPTransportEnvVerifyServerCert(t *testing.T) {
-	var h recordingHandler
-	server := httptest.NewTLSServer(&h)
-	defer server.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_URL", server.URL)()
-	defer patchEnv("ELASTIC_APM_VERIFY_SERVER_CERT", "false")()
-
-	transport, err := transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.NoError(t, err)
-
-	assert.NotNil(t, transport.Client)
-	assert.IsType(t, &http.Transport{}, transport.Client.Transport)
-	httpTransport := transport.Client.Transport.(*http.Transport)
-	assert.NotNil(t, httpTransport.TLSClientConfig)
-	assert.True(t, httpTransport.TLSClientConfig.InsecureSkipVerify)
-
-	err = transport.SendStream(context.Background(), strings.NewReader(""))
-	assert.NoError(t, err)
-}
-
 func TestHTTPError(t *testing.T) {
 	h := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 		http.Error(w, "error-message", http.StatusInternalServerError)
@@ -273,37 +252,6 @@ func TestHTTPTransportServerTimeout(t *testing.T) {
 	})
 }
 
-func TestHTTPTransportServerFailover(t *testing.T) {
-	defer patchEnv("ELASTIC_APM_VERIFY_SERVER_CERT", "false")()
-
-	var hosts []string
-	errorHandler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		hosts = append(hosts, req.Host)
-		http.Error(w, "error-message", http.StatusInternalServerError)
-	})
-	server1 := httptest.NewServer(errorHandler)
-	defer server1.Close()
-	server2 := httptest.NewTLSServer(errorHandler)
-	defer server2.Close()
-
-	transport, err := transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	require.NoError(t, err)
-	transport.SetServerURL(mustParseURL(server1.URL), mustParseURL(server2.URL))
-
-	for i := 0; i < 4; i++ {
-		err := transport.SendStream(context.Background(), strings.NewReader(""))
-		assert.EqualError(t, err, "request failed with 500 Internal Server Error: error-message")
-	}
-	assert.Len(t, hosts, 4)
-
-	// Each time SendStream returns an error, the transport should switch
-	// to the next URL in the list. The list is shuffled so we only compare
-	// the output values to each other, rather than to the original input.
-	assert.NotEqual(t, hosts[0], hosts[1])
-	assert.Equal(t, hosts[0], hosts[2])
-	assert.Equal(t, hosts[1], hosts[3])
-}
-
 func TestHTTPTransportV2NotFound(t *testing.T) {
 	server := httptest.NewServer(http.NotFoundHandler())
 	defer server.Close()
@@ -314,132 +262,6 @@ func TestHTTPTransportV2NotFound(t *testing.T) {
 
 	err = transport.SendStream(context.Background(), strings.NewReader(""))
 	assert.EqualError(t, err, fmt.Sprintf("request failed with 404 Not Found: %s/intake/v2/events not found (requires APM Server 6.5.0 or newer)", server.URL))
-}
-
-func TestHTTPTransportCACert(t *testing.T) {
-	var h recordingHandler
-	server := httptest.NewUnstartedServer(&h)
-	server.Config.ErrorLog = log.New(ioutil.Discard, "", 0)
-	server.StartTLS()
-	defer server.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_URL", server.URL)()
-
-	p := strings.NewReader("")
-
-	// SendStream should fail, because we haven't told the client about
-	// the server certificate, nor disabled certificate verification.
-	trans, err := transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.NoError(t, err)
-	assert.NotNil(t, trans)
-	err = trans.SendStream(context.Background(), p)
-	assert.Error(t, err)
-
-	// Set the env var to a file that doesn't exist, should get an error
-	defer patchEnv("ELASTIC_APM_SERVER_CA_CERT_FILE", "./testdata/file_that_doesnt_exist.pem")()
-	trans, err = transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.Error(t, err)
-	assert.Nil(t, trans)
-
-	// Set the env var to a file that has no cert, should get an error
-	f, err := ioutil.TempFile("", "apm-test-1")
-	require.NoError(t, err)
-	defer os.Remove(f.Name())
-	defer f.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_CA_CERT_FILE", f.Name())()
-	trans, err = transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.Error(t, err)
-	assert.Nil(t, trans)
-
-	// Set a certificate that doesn't match, SendStream should still fail
-	defer patchEnv("ELASTIC_APM_SERVER_CA_CERT_FILE", "./testdata/cert.pem")()
-	trans, err = transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.NoError(t, err)
-	assert.NotNil(t, trans)
-	err = trans.SendStream(context.Background(), p)
-	assert.Error(t, err)
-
-	f, err = ioutil.TempFile("", "apm-test-2")
-	require.NoError(t, err)
-	defer os.Remove(f.Name())
-	defer f.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_CA_CERT_FILE", f.Name())()
-
-	err = pem.Encode(f, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: server.TLS.Certificates[0].Certificate[0],
-	})
-	require.NoError(t, err)
-
-	trans, err = transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.NoError(t, err)
-	assert.NotNil(t, trans)
-	err = trans.SendStream(context.Background(), p)
-	assert.NoError(t, err)
-}
-
-func TestHTTPTransportServerCert(t *testing.T) {
-	var h recordingHandler
-	server := httptest.NewUnstartedServer(&h)
-	server.Config.ErrorLog = log.New(ioutil.Discard, "", 0)
-	server.StartTLS()
-	defer server.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_URL", server.URL)()
-
-	p := strings.NewReader("")
-
-	newTransport := func() *transport.HTTPTransport {
-		transport, err := transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-		require.NoError(t, err)
-		return transport
-	}
-
-	// SendStream should fail, because we haven't told the client about
-	// the server certificate, nor disabled certificate verification.
-	transport := newTransport()
-	err := transport.SendStream(context.Background(), p)
-	assert.Error(t, err)
-
-	// Set a certificate that doesn't match, SendStream should still fail.
-	defer patchEnv("ELASTIC_APM_SERVER_CERT", "./testdata/cert.pem")()
-	transport = newTransport()
-	err = transport.SendStream(context.Background(), p)
-	assert.Error(t, err)
-
-	f, err := ioutil.TempFile("", "apm-test")
-	require.NoError(t, err)
-	defer os.Remove(f.Name())
-	defer f.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_CERT", f.Name())()
-
-	// Reconfigure the transport so that it knows about the
-	// server certificate. We avoid using server.Client here, as
-	// it is not available in older versions of Go.
-	err = pem.Encode(f, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: server.TLS.Certificates[0].Certificate[0],
-	})
-	require.NoError(t, err)
-
-	transport = newTransport()
-	err = transport.SendStream(context.Background(), p)
-	assert.NoError(t, err)
-}
-
-func TestHTTPTransportServerCertInvalid(t *testing.T) {
-	f, err := ioutil.TempFile("", "apm-test")
-	require.NoError(t, err)
-	defer os.Remove(f.Name())
-	defer f.Close()
-	defer patchEnv("ELASTIC_APM_SERVER_CERT", f.Name())()
-
-	fmt.Fprintln(f, `
------BEGIN GARBAGE-----
-garbage
------END GARBAGE-----
-`[1:])
-
-	_, err = transport.NewHTTPTransport(transport.HTTPTransportOptions{})
-	assert.EqualError(t, err, fmt.Sprintf("failed to load certificate from %s: missing or invalid certificate", f.Name()))
 }
 
 func TestHTTPTransportWatchConfig(t *testing.T) {
